// src/models/Docentes.ts
export interface IDocentes {
    idDocente: number;
    codigo: string;
    documentoIdentidad: string;
    nombres: string;
    apellidoPaterno: string;
    apellidoMaterno: string;
    correoInstitucional: string;
    fechaNacimiento: Date;
    sexo: string;
    direccion:string;
    especialidad:string;
    estadoAuditoria: string;
    fechaCreacion: Date;
    fechaModificacion: Date;
}